% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{ci}
\alias{ci}
\alias{ci.default}
\alias{ci.numeric}
\title{Return confidence intervals of a variable}
\usage{
\method{ci}{numeric}(x, level = 0.95, na.rm = TRUE, df = Inf)

\method{ci}{default}(x, level = 0.95, na.rm = TRUE)

ci(x, ...)
}
\arguments{
\item{x}{A vector}

\item{level}{Confidence level}

\item{na.rm}{Logical. Should missing values be removed?}

\item{df}{Degrees of freedom. By default this is \code{Inf} which is equivalent to assuming a normal distribution.}
}
\value{
A length-two vector giving a confidence interval.
If \code{x} is numeric, the confidence interval around the mean is calculated using the t distribution.

If \code{x} is logical, or a factor with two levels, the binomial confidence interval is calculated for the
proportion of \code{TRUE} cases, or cases on the first level, using \code{\link{prop.test}}.
}
\description{
Return confidence intervals of a variable
}
\examples{
ci(rnorm(10), 0.95)
a <- rnorm(10)
ci(a, 0.99)
ci(a, 0.99, df = 10)
}

