% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_style.R
\docType{data}
\name{ff_colors}
\alias{ff_colors}
\title{Forsyth Futures' color palette}
\format{
An object of class \code{character} of length 43.
}
\usage{
ff_colors
}
\description{
Vector of Forsyth Futures' color palette
}
\keyword{ff_colors}
