% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_plot.R
\name{ff_ggplot}
\alias{ff_ggplot}
\title{Create Forsyth Futures' themed ggplot graph}
\usage{
ff_ggplot(plot_name, source_name)
}
\arguments{
\item{plot_name}{The object name of the plot you have created that you want to format}

\item{source_name}{The text you want to come after the text 'Source:' in the bottom left hand side of your plot}
}
\description{
This function allows you to create a static ggplot graph with Forsyth Futures' theming.
}
\examples{
p <- ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg)) +
  ggplot2::geom_point() +
  ggplot2::labs(title = "Fuel economy declines as weight increases",
                subtitle = "Fuel economy declines as weight increases")
ff_ggplot(plot_name = p, source = "The source for my data")
}
\keyword{ff_ggplot}
