% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_plot.R
\name{ff_ggplotly}
\alias{ff_ggplotly}
\title{Create Forsyth Futures' themed plotly graph}
\usage{
ff_ggplotly(plot_name, source_name)
}
\arguments{
\item{plot_name}{The object name of the plot you have created that you want to format}

\item{source_name}{The text you want to come after the text 'Source:' in the bottom left hand side of your plot}
}
\description{
This function allows you to convert a ggplot graph into an interactive plotly graphic with Forsyth Futures' theming.
}
\examples{
p <- ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg)) +
  ggplot2::geom_point(
   ggplot2::aes(
    text = paste0(
      "wt: ", wt, "<br>",
      "mpg: ", mpg, "<br>")
    )
  ) +
  ggplot2::labs(title = "Fuel economy declines as weight increases",
                subtitle = "Fuel economy declines as weight increases")
ff_ggplotly(plot_name = p, source_name = "The source for my data")
}
\keyword{ff_ggplotly}
