% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_style.R
\name{ff_style}
\alias{ff_style}
\title{Add Forsyth Futures' theme to ggplot graph}
\usage{
ff_style()
}
\description{
This function allows you to add the Forsyth Futures' theme to your ggplot graphics.
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg)) +
 ggplot2::geom_point() +
 ggplot2::labs(title = "Fuel economy declines as weight increases",
               subtitle = "Fuel economy declines as weight increases") +
 ff_style()
}
\keyword{ff_style}
