% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ffplot.R
\name{fftable.default}
\alias{fftable}
\alias{fftable.default}
\alias{fftable.formula}
\title{Create a table from a formula}
\usage{
\method{fftable}{default}(data, ...)

\method{fftable}{formula}(formula, data = parent.frame(), subset = NULL)
}
\arguments{
\item{data}{a data frame (or list or environment) containing the variables referred to in \code{formula}.}

\item{formula}{a formula object with exactly one term on the LHS and RHS.}

\item{subset}{an optional vector specifying a subset of \code{data}.}
}
\value{
A data frame like that from \code{\link{aggregate}}.
}
\description{
\code{fftable} splits \code{data} by unique values of \code{formula}'s right hand side, and
evaluates the left hand side for each subset.
}
\details{
The default method simply passes its arguments on to \code{fftable.formula}, taking the first argument as
\code{data}. This plays nicely with \code{\link[dplyr]{dplyr-package}}.
}
\examples{
fftable(mean(mpg) ~ gear, mtcars)
fftable(range(mpg) ~ gear, mtcars)
fftable(mpg ~ gear, mtcars)
fftable(cbind(range(mpg), quantile(mpg, c(0.25, 0.75))) ~ gear, mtcars)
\dontrun{
library(dplyr)
mtcars \%>\% fftable(range(mpg) ~ gear)
}
}

