% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points-total.R
\name{points_total}
\alias{points_total}
\title{Add together two numbers}
\usage{
points_total(data)
}
\arguments{
\item{data}{The data frame from \code{\link[fflr:league_standings]{fflr::league_standings()}}.}
}
\value{
A ggplot bar chart object.
}
\description{
Add together two numbers
}
\examples{
points_total(fflr::league_standings(252353))
}
