% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{prop}
\alias{prop}
\title{Calculates the proportion of \code{TRUE} cases or cases matching the first level}
\usage{
prop(x)
}
\arguments{
\item{x}{a logical vector or a factor with only 2 levels}
}
\value{
A numeric
}
\description{
Unlike \code{mean}, if used inside \code{\link{ffplot}} this will automatically use a barplot
}

