% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standings-change.R
\name{standings_change}
\alias{standings_change}
\title{Plot change in standing projections}
\usage{
standings_change(data, filename = NULL)
}
\arguments{
\item{data}{The data frame from \code{\link[fflr:league_standings]{fflr::league_standings()}}.}

\item{filename}{File name to create on disk. Passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}},
with a height of 4 inches, width of 7, and a DPI of 320. If \code{NULL}, no file
is created.}
}
\value{
A ggplot bar chart object.
}
\description{
Plot change in standing projections
}
\examples{
tmp <- tempfile()
standings_change(fflr::league_standings(252353), filename = tmp)
}
