% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team-points.R
\name{team_points}
\alias{team_points}
\title{Add together two numbers}
\usage{
team_points(data)
}
\arguments{
\item{data}{The data frame from \code{\link[fflr:league_standings]{fflr::league_standings()}}.}
}
\value{
A ggplot bar chart object.
}
\description{
Add together two numbers
}
\examples{
library(fflr)
team_points(league_standings(252353))
}
