% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly-scores.R
\name{weekly_power}
\alias{weekly_power}
\title{Plot bar chart of weekly power wins}
\usage{
weekly_power(data)
}
\arguments{
\item{data}{The data frame from \code{\link[fflr:weekly_matchups]{fflr::weekly_matchups()}}.}
}
\value{
A ggplot bar chart object.
}
\description{
Plot bar chart of weekly power wins
}
\examples{
weekly_power(fflr::weekly_matchups(252353))
}
