% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly-scores.R
\name{weekly_scores}
\alias{weekly_scores}
\title{Plot bar chart of weekly scores}
\usage{
weekly_scores(data)
}
\arguments{
\item{data}{The data frame from \code{\link[fflr:weekly_matchups]{fflr::weekly_matchups()}}.}
}
\value{
A ggplot bar chart object.
}
\description{
Plot bar chart of weekly scores
}
\examples{
weekly_scores(fflr::weekly_matchups(252353))
}
