% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ffplot.R
\name{ci}
\alias{ci}
\title{Return confidence intervals of a variable}
\usage{
ci(x, level = 0.95, na.rm = TRUE, df = Inf)
}
\arguments{
\item{x}{A numeric vector}

\item{level}{Confidence level}

\item{na.rm}{Logical. Should missing values be removed?}

\item{df}{Degrees of freedom. By default this is \code{Inf} which is equivalent to assuming a normal distribution.}
}
\value{
A length-two vector giving the confidence interval around the mean.
}
\description{
Return confidence intervals of a variable
}
\examples{
ci(rnorm(10), 0.95)
a <- rnorm(10)
ci(a, 0.99)
ci(a, 0.99, df = 10)
}

